up merge

+++ Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/model/items/enchant/EnchantRateItem.java
@@ -1,19 +1,3 @@
-/*
- * This file is part of the L2J Mobius project.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
 package org.l2j.gameserver.model.items.enchant;
 
 import org.l2j.gameserver.model.items.ItemTemplate;
@@ -24,7 +8,7 @@ import org.l2j.gameserver.model.items.ItemTemplate;
 public final class EnchantRateItem {
     private final String _name;
     private int _itemId;
-    private int _slot;
+    private long _slot;
     private Boolean _isMagicWeapon = null;
 
     public EnchantRateItem(String name) {
